Cleaned up repo with `gofumpt` Implemented minor support for `darwin` Implemented automated builds Improved `README`
