+++ .github/workflows/build.yml
@@ -1,0 +1,21 @@
+name: Build
+on: pull_request
+
+jobs:
+  build:
+    name: Build and Release
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout Repo
+        uses: actions/checkout@master
+
+      - name: Setup Go Environment
+        uses: actions/setup-go@v2
+        with:
+          go-version: '^1.18.0'
+
+      - name: Build Binaries
+        run: |
+          go install github.com/fyne-io/fyne-cross@latest
+          fyne-cross linux -arch=amd64,arm64
+          fyne-cross windows -arch=amd64
