+++ globals/global_variables.go
@@ -7,22 +7,28 @@ import (
 	"fyne.io/fyne/v2/app"
 )
 
-var Original_images []string
-var Original_paths []string
-var Resized_images []string
+var (
+	Original_images []string
+	Original_paths  []string
+	Resized_images  []string
+)
 
 const SYS_OS = runtime.GOOS
 
-var MyApp = app.NewWithID("walldo")
-var Window = MyApp.NewWindow("Walldo in go")
+var (
+	MyApp  = app.NewWithID("walldo")
+	Window = MyApp.NewWindow("Walldo in go")
+)
 
 // configuracion
 var GridSize = MyApp.Preferences().StringWithFallback("GridSize", "default")
 var GridTitles = MyApp.Preferences().StringWithFallback("GridTitles", "Borderless")
 
-var LayoutStyle = MyApp.Preferences().StringWithFallback("Layout", "Grid")
-var FillStrategy = MyApp.Preferences().StringWithFallback("FillStrategy", "Zoom Fill")
- 
+var (
+	LayoutStyle  = MyApp.Preferences().StringWithFallback("Layout", "Grid")
+	FillStrategy = MyApp.Preferences().StringWithFallback("FillStrategy", "Zoom Fill")
+)
+
 // archivos de config
 var ConfigDir, _ = os.UserHomeDir() // home del usuario
 var ConfigPath = ConfigDir          // path de las configuraciones
@@ -31,7 +37,7 @@ var ConfigPath = ConfigDir          // path de las configuraciones
 // ~/.config/walldo/config.json (unix)
 // ~/AppData/Local/walldo/config.json (windows)
 func SetGlobalValues() {
-    os.Setenv("FYNE_THEME", "dark")
+	os.Setenv("FYNE_THEME", "dark")
 
 	switch SYS_OS {
 	case "windows":
@@ -44,4 +50,3 @@ func SetGlobalValues() {
 		ConfigPath += "/.config/walldo/"
 	}
 }
-
